assets/handlers/python/login/pyproject.toml
[build-system]
requires = ["setuptools>=63.0"]
build-backend = "setuptools.build_meta"

[project]
name = "login"
description = "Login a user using username/email and password authentication."
version = "1.0.0"
requires-python = ">=3.9"
dependencies = [
  "argon2-cffi>=23.1.0,<23.2.0",
  "motor>=3.2.2,<3.3.2",
  "pydantic>=2.5.2,<2.6.0",
]
